requirements.txt
PyPDF2==3.0.0
aspose-words==22.12.0